+++ changes.md
@@ -1,6 +1,7 @@
 Change Log
 ==========
 
+
 ### changes in 1.0
 
 * remove all circe dependencies
@@ -10,3 +11,6 @@ Change Log
 * added custom properties
 * added temporary scalastix-0.3-SNAPSHOT.jar
 * added the missing hashes of ExternalReference
+* remove description from ObservablesMaker
+* added custom (for custom properties x_...) to ObservablesMaker
+* update to scalastix-05